requirements.txt
GDAL==3.6.2
matplotlib==3.10.0
numpy==2.2.1
Pillow==9.4.0
Pillow==11.1.0
python-dotenv==1.0.1
segmentation_models_pytorch==0.4.0
torch==2.5.1
torchmetrics==1.6.1
torchvision==0.20.1
tqdm==4.67.1
